+++ toll_fee_calc_app/src/main/java/com/evolve_technology/calculator/controller/TollFeeController.java
@@ -20,19 +20,19 @@ import com.evolve_technology.calculator.service.TollFeeService;
 public class TollFeeController {
 	
 	@Autowired
-	TollFeeService tollFreeService;
+	TollFeeService tollFeeService;
 	
 	@GetMapping("/vehicles/{name}/dates/{date}")
 	public ResponseEntity<TollFee> getTollFee(@PathVariable("name") String name,@PathVariable("date") 
 	@Pattern(regexp = "^((20)[0-9][0-9])-(0[1-9]|1[012])-(0[1-9]|[12][0-9]|3[01])T(00|0[0-9]|1[0-9]|2[0-3]):(0[0-9]|[0-5][0-9]):(0[0-9]|[0-5][0-9])$") 
 	String date) {
-		TollFee tollFee= tollFreeService.getTollFee(date, name);
+		TollFee tollFee= tollFeeService.getTollFee(date, name);
 		return new ResponseEntity<>(tollFee,HttpStatus.OK);
 	}
 
 	@GetMapping("/records")
 	public Map<String,Map<String,Map<Integer,Integer>>> getTollFeeHistoricalRecords(){
-		return tollFreeService.getHistoricalTollRecords();
+		return tollFeeService.getHistoricalTollRecords();
 	}
 	
 }
