+++ toll_fee_calc_app/src/main/java/com/evolve_technology/calculator/properties/TollConfiguration.java
@@ -2,6 +2,7 @@ package com.evolve_technology.calculator.properties;
 
 import java.util.List;
 
+import org.springframework.beans.factory.annotation.Value;
 import org.springframework.boot.context.properties.ConfigurationProperties;
 import org.springframework.stereotype.Component;
 
@@ -12,6 +13,7 @@ import lombok.Data;
 @ConfigurationProperties(prefix="tollfree")
 public class TollConfiguration {
 	
+	@Value("${year}")
 	private String year;
 	private List<String> dates;
 	private List<String> vehicles;
