+++ toll_fee_calc_app/src/main/java/com/evolve_technology/calculator/service/impl/TollFreeDatesServiceImpl.java
@@ -2,7 +2,6 @@ package com.evolve_technology.calculator.service.impl;
 
 import java.time.DayOfWeek;
 import java.time.LocalDate;
-import java.time.Month;
 import java.util.List;
 
 import org.apache.logging.log4j.LogManager;
