+++ toll_fee_calc_app/src/main/resources/application.properties
@@ -1,7 +1,7 @@
 server.servlet.context-path=/v1/api
 
-# Toll Free Year
-tollfree.year=2013
+# Year
+year=2013
 
 # Toll Free Dates
 tollfree.dates=2013-01-01 , 2013-03-28 , 2013-03-29 , 2013-04-01 ,  2013-04-30 , 2013-05-01 , 2013-05-08 , 2013-05-09 , 2013-06-05 , 2013-06-06 , 2013-06-21 , 2013-11-01 , 2013-12-24 , 2013-12-25 , 2013-12-26 , 2013-12-31
